+++ packages/forms/src/index.ts
@@ -14,3 +14,8 @@ export * from "./components/ui/text-input/SimpleTextInput";
 export * from "./components/ui/text-input/SimpleTextInputTheme";
 export * from "./components/ui/text-input/StandardTextInput";
 export * from "./components/ui/text-input/StandardTextInputTheme";
+export * from "./components/ui/text-input/duration-text-input/DurationTextInput";
+export * from "./components/ui/types";
+export * from "./utils/DurationHelper";
+export * from "./utils/TimeStringHelper";
+export * from "./utils/types/HoursAndMinutes";