+++ packages/forms/src/utils/types/HoursAndMinutes.ts
@@ -1,0 +1,4 @@
+export interface HourAndMinutes {
+  hours: number;
+  minutes: number;
+}